Move GooeyParamWidgetMixin into its own module
Got crowded in param_widgets.py
